+++ shared/domain/src/commonMain/kotlin/com/sedsoftware/tackle/domain/model/Instance.kt
@@ -40,16 +40,16 @@ data class Instance(
         )
 
         data class Polls(
-            val maxOptions: Long = 0L,
-            val maxCharactersPerOption: Long = 0L,
+            val maxOptions: Int = 0,
+            val maxCharactersPerOption: Int = 0,
             val minExpiration: Long = 0L,
             val maxExpiration: Long = 0L,
         )
 
         class Statuses(
-            val maxCharacters: Long = 0L,
-            val maxMediaAttachments: Long = 0L,
-            val charactersReservedPerUrl: Long = 0L,
+            val maxCharacters: Int = 0,
+            val maxMediaAttachments: Int = 0,
+            val charactersReservedPerUrl: Int = 0,
         )
     }
